Add middleware to decode inference request inputs
